fix: 리뷰 msw 분기처리
