+++ src/mocks/handlers/review/index.ts
@@ -1,11 +1,5 @@
 import { createReview } from './createReview';
 import { popularReview } from './popularReview';
 import { reviewList } from './reviewList';
-import { travelDetailReview } from './travelDetailReview';
 
-export const review = [
-  createReview,
-  ...popularReview,
-  ...reviewList,
-  ...travelDetailReview,
-];
+export const review = [createReview, ...popularReview, reviewList];
