+++ src/mocks/handlers/review/mockTravelDetailReview.ts
@@ -1,0 +1,6 @@
+import { HttpResponse } from 'msw';
+import reviewList from '@/mocks/data/review/travelReviewList.json';
+
+export const mockTravelDetailReview = () => {
+  return HttpResponse.json(reviewList);
+};
