+++ src/mocks/handlers/review/reviewList.ts
@@ -1,47 +1,16 @@
-import { http, HttpResponse } from 'msw';
-import reviewListMock from '@/mocks/data/review/reviewInfinityListMock.json';
-import reviewListMock2 from '@/mocks/data/review/reviewInfinityListMock2.json';
+import { http } from 'msw';
 
-export const reviewList = [
-  http.get(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/reviews`,
-    async ({ request }) => {
-      const url = new URL(request.url);
-      const requestedPage = parseInt(url.searchParams.get('page') || '1', 10);
-      const limit = parseInt(url.searchParams.get('limit') || '12', 10);
-      const sortOrder = url.searchParams.get('sortBy') || 'createdAt';
+import { mockTravelDetailReview } from './mockTravelDetailReview';
+import { mockReviewList } from './mockReviewList';
 
-      const pageSize = limit; // 페이지 크기
-      const startIndex = (requestedPage - 1) * pageSize; // 시작 인덱스
-      const endIndex = startIndex + pageSize; // 끝 인덱스
-      const reviewsToSend = reviewListMock.reviews.slice(startIndex, endIndex);
-      const reviewsToSend2 = reviewListMock2.reviews.slice(
-        startIndex,
-        endIndex,
-      );
-
-      if (sortOrder === 'createdAt') {
-        const response = {
-          currentPage: requestedPage,
-          size: reviewsToSend.length, // 현재 페이지의 리뷰 수
-          isFirst: requestedPage === 1,
-          isLast: requestedPage * pageSize >= reviewListMock.reviews.length,
-          reviews: reviewsToSend, // 제한된 리뷰 반환
-        };
-        return HttpResponse.json(response);
-      }
-      if (sortOrder === 'popular') {
-        const response = {
-          currentPage: requestedPage,
-          size: reviewsToSend.length, // 현재 페이지의 리뷰 수
-          isFirst: requestedPage === 1,
-          isLast: requestedPage * pageSize >= reviewListMock.reviews.length,
-          reviews: reviewsToSend2, // 제한된 리뷰 반환
-        };
-        return HttpResponse.json(response);
-      }
-
-      return HttpResponse.error();
-    },
-  ),
-];
+export const reviewList = http.get(
+  `${process.env.NEXT_PUBLIC_BASE_URL}/reviews`,
+  async ({ request }) => {
+    const url = new URL(request.url);
+    const id = url.searchParams.get('id');
+    if (id) {
+      return mockTravelDetailReview();
+    }
+    return mockReviewList(url);
+  },
+);
